Version 48
Note:  In the revision notes, it lists changes in the export coefficient input file(s).  This part of the code is not completely operational in this revision,
so these changes will not be reflected in the example input datasets or in the modular spreadsheet.  This work in progress will be functional in the next few revisions.

--The executables for Revision 48 (64 bit debug/release) have been added to this revision.

Many routines have changed due to the rest of the IMPLICIT NONE addition to the code.  There have also been several routines added to the code due to the export coefficient changes that are taking place.  (see revision notes).